+++ handler.go
@@ -14,6 +14,7 @@ type handler struct {
 	do     reflect.Value // do should return response for HTTP writer
 	method string
 
+	guards       []*context.UserContext
 	userContexts []*context.UserContext
 
 	matchedPath      string
@@ -28,18 +29,27 @@ func newHandler(do reflect.Value) *handler {
 }
 
 func newErrorHandler(code int, content string) *handler {
-	h := newHandler(reflect.ValueOf(func() *response.Response {
+	return newHandler(reflect.ValueOf(func() *response.Response {
 		return response.New(content).Status(code)
 	}))
-	return h
 }
 
 func (h *handler) handle(reqURL []string, r *http.Request) *response.Response {
-	ctx, err := h.getUserContexts(reqURL, r)
+	ctx, err := h.getContexts(reqURL, r)
 	if err != nil {
 		return response.New(err.Error()).
 			Status(http.StatusBadRequest)
 	}
+
+	if err := h.verify(reqURL, r); err != nil {
+		if err, ok := err.(*VerifyError); ok {
+			return response.New(err.Error()).
+				Status(err.Status())
+		}
+		return response.New(err.Error()).
+			Status(http.StatusInternalServerError)
+	}
+
 	resp := h.do.Call(
 		ctx,
 	)[0].Interface()
@@ -52,6 +62,38 @@ func (h *handler) handle(reqURL []string, r *http.Request) *response.Response {
 	}
 }
 
+func (h *handler) Guard(guard Guard) *handler {
+	if h.guards == nil {
+		h.guards = make([]*context.UserContext, 0)
+	}
+	contextT := reflect.TypeOf(guard).Elem()
+	h.guards = append(h.guards,
+		context.
+			NewUserContext().
+			CacheParamsOffset(contextT, h.routes),
+	)
+	return h
+}
+
+func (h *handler) verify(reqURL []string, r *http.Request) error {
+	// no guards
+	if h.guards == nil {
+		return nil
+	}
+	ctx, err := h.getGuards(reqURL, r)
+	if err != nil {
+		return err
+	}
+	for _, c := range ctx {
+		// without check since we already do the static checking by method signature
+		guard := c.Interface().(Guard)
+		if err := guard.VerifyRequest(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func (h *handler) addMatchedPathValueIntoContext(paths ...string) {
 	path := bytes.NewBuffer([]byte(``))
 	for _, v := range paths {
@@ -61,11 +103,19 @@ func (h *handler) addMatchedPathValueIntoContext(paths ...string) {
 	h.matchedPath = path.String()[:path.Len()-1]
 }
 
-func (h *handler) getUserContexts(reqURL []string, req *http.Request) ([]reflect.Value, error) {
-	userContexts := make([]reflect.Value, h.do.Type().NumIn())
+func (h *handler) getContexts(reqURL []string, req *http.Request) ([]reflect.Value, error) {
+	return h.fillByCachedUserContexts(h.userContexts, reqURL, req)
+}
+
+func (h *handler) getGuards(reqURL []string, req *http.Request) ([]reflect.Value, error) {
+	return h.fillByCachedUserContexts(h.guards, reqURL, req)
+}
+
+func (h *handler) fillByCachedUserContexts(contexts []*context.UserContext, reqURL []string, req *http.Request) ([]reflect.Value, error) {
+	userContexts := make([]reflect.Value, len(contexts))
 
 	req.ParseForm()
-	for i, userContext := range h.userContexts {
+	for i, userContext := range contexts {
 		basicChain := []filler{
 			newRouteFiller(
 				h.routes,